# backend/WandelApp.api/WandelApp.api2/Models/Games/PutGameModel.cs
using System;
using System.Collections.Generic;

namespace WandelApp.api.Games
{
    public class PutGameModel : BaseGameModel
    {
        public ICollection<Guid> UserGames { get; set; }
    }
}