# backend/WandelApp.api/WandelApp.api2/Models/UserGames/PutUserGameModel.cs
using System.ComponentModel.DataAnnotations;

namespace WandelApp.api.UserGames
{
    public class PutUserGameModel : BaseUserGameModel
    {
        [Required]
        public int CountPlayed { get; set; }
    }
}